requirements.txt
schedule==0.6.0
requests==2.22.0
Pillow==7.0.0
